Integrate lots of new SDK functionality
This commit eliminates a lot of the hard-coded simulation-only values, using the SDK to get execution status, etc.

It also adds a method to parse some values of interest out of the nepi-bot config file. That functionality may eventually get moved to the SDK, but it is not clear at this point that the data acquired is of general interest.